split out constraints
