+++ lib/core/src/Cardano/Wallet/Primitive/Types/Tx/Tx.hs
@@ -38,15 +38,6 @@ module Cardano.Wallet.Primitive.Types.Tx.Tx
     , txOutSubtractCoin
     , txScriptInvalid
 
-    -- * Constraints
-    , TxConstraints (..)
-    , txOutputCoinCost
-    , txOutputCoinSize
-    , txOutputHasValidSize
-    , txOutputHasValidTokenQuantities
-    , TxSize (..)
-    , txSizeDistance
-
     -- * Queries
     , txAssetIds
     , txOutAssetIds
@@ -78,9 +69,7 @@ import Cardano.Wallet.Primitive.Types.RewardAccount
 import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle (..) )
 import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId, Lexicographic (..), TokenMap )
-import Cardano.Wallet.Primitive.Types.TokenQuantity
-    ( TokenQuantity (..) )
+    ( AssetId, Lexicographic (..) )
 import Cardano.Wallet.Primitive.Types.Tx.CBOR
     ( TxCBOR )
 import Control.DeepSeq
@@ -113,10 +102,6 @@ import Fmt
     )
 import GHC.Generics
     ( Generic )
-import Numeric.Natural
-    ( Natural )
-import Quiet
-    ( Quiet (..) )
 
 import qualified Cardano.Wallet.Primitive.Types.Coin as Coin
 import qualified Cardano.Wallet.Primitive.Types.TokenBundle as TokenBundle
@@ -380,91 +365,6 @@ data TokenBundleSizeAssessment
     -- ^ Indicates that the size of a token bundle exceeds the maximum size
     -- that can be included in a transaction output.
     deriving (Eq, Generic, Show)
-
---------------------------------------------------------------------------------
--- Constraints
---------------------------------------------------------------------------------
-
--- | Provides an abstract cost and size model for transactions.
---
--- This allows parts of a transaction to be costed (or sized) individually,
--- without having to compute the cost (or size) of an entire transaction.
---
--- Note that the following functions assume one witness is required per input:
---
--- - 'txInputCost'
--- - 'txInputSize'
---
--- This will lead to slight overestimation in the case of UTxOs that share the
--- same payment key.
---
-data TxConstraints = TxConstraints
-    { txBaseCost :: Coin
-      -- ^ The constant cost of an empty transaction.
-    , txBaseSize :: TxSize
-      -- ^ The constant size of an empty transaction.
-    , txInputCost :: Coin
-      -- ^ The constant cost of a transaction input, assuming one witness is
-      -- required per input.
-    , txInputSize :: TxSize
-      -- ^ The constant size of a transaction input, assuming one witness is
-      -- required per input.
-    , txOutputCost :: TokenBundle -> Coin
-      -- ^ The variable cost of a transaction output.
-    , txOutputSize :: TokenBundle -> TxSize
-      -- ^ The variable size of a transaction output.
-    , txOutputMaximumSize :: TxSize
-      -- ^ The maximum size of a transaction output.
-    , txOutputMaximumTokenQuantity :: TokenQuantity
-      -- ^ The maximum token quantity that can appear in a transaction output.
-    , txOutputMinimumAdaQuantity :: Address -> TokenMap -> Coin
-      -- ^ The variable minimum ada quantity of a transaction output.
-    , txOutputBelowMinimumAdaQuantity :: Address -> TokenBundle -> Bool
-      -- ^ Returns 'True' if the given 'TokenBundle' has a 'Coin' value that is
-      -- below the minimum required.
-    , txRewardWithdrawalCost :: Coin -> Coin
-      -- ^ The variable cost of a reward withdrawal.
-    , txRewardWithdrawalSize :: Coin -> TxSize
-      -- ^ The variable size of a reward withdrawal.
-    , txMaximumSize :: TxSize
-      -- ^ The maximum size of a transaction.
-    }
-    deriving Generic
-
-txOutputCoinCost :: TxConstraints -> Coin -> Coin
-txOutputCoinCost constraints = txOutputCost constraints . TokenBundle.fromCoin
-
-txOutputCoinSize :: TxConstraints -> Coin -> TxSize
-txOutputCoinSize constraints = txOutputSize constraints . TokenBundle.fromCoin
-
-txOutputHasValidSize :: TxConstraints -> TokenBundle -> Bool
-txOutputHasValidSize constraints b =
-    txOutputSize constraints b <= txOutputMaximumSize constraints
-
-txOutputHasValidTokenQuantities :: TxConstraints -> TokenMap -> Bool
-txOutputHasValidTokenQuantities constraints m =
-    TokenMap.maximumQuantity m <= txOutputMaximumTokenQuantity constraints
-
--- | The size of a transaction, or part of a transaction, in bytes.
---
-newtype TxSize = TxSize { unTxSize :: Natural }
-    deriving stock (Eq, Ord, Generic)
-    deriving Show via (Quiet TxSize)
-
-instance NFData TxSize
-
-instance Semigroup TxSize where
-    TxSize a <> TxSize b = TxSize (a + b)
-
-instance Monoid TxSize where
-    mempty = TxSize 0
-
--- | Computes the absolute distance between two transaction size quantities.
---
-txSizeDistance :: TxSize -> TxSize -> TxSize
-txSizeDistance (TxSize a) (TxSize b)
-    | a >= b    = TxSize (a - b)
-    | otherwise = TxSize (b - a)
 
 --------------------------------------------------------------------------------
 -- Queries
